pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gapverify"
version = "0.1.0"
description = "GapVerify: Latent Discrepancy Control for Retrieval-Grounded Verification"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "GapVerify Team" }
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "torch>=2.1",
  "transformers>=4.40",
  "accelerate>=0.30",
  "datasets>=2.18",
  "sentence-transformers>=2.6",
  "faiss-cpu>=1.7.4",
  "scikit-learn>=1.3",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "ruff>=0.15.6"
]

[tool.setuptools]
packages = ["gapverify"]

[tool.ruff]
line-length = 100
target-version = "py310"
